Add core versions of suspending and chained commands
